setting up automated pipeline

+++ src/02_dataset_download_ibtracs.py
@@ -18,10 +18,11 @@ def main():
     ibtracs_path = os.path.join(DATA_PATH, "ibtracs.since1980.list.v04r01.csv")
     yprov4ml.log_param("ibtracs_url", ibtracs_url)
 
-    response = requests.get(ibtracs_url)
-    file = open(ibtracs_path, "wb")
-    for data in tqdm(response.iter_content(chunk_size=1024), unit="kB"):
-        file.write(data)
+    if not os.path.exists(ibtracs_path): 
+        response = requests.get(ibtracs_url)
+        file = open(ibtracs_path, "wb")
+        for data in tqdm(response.iter_content(chunk_size=1024), unit="kB"):
+            file.write(data)
 
     yprov4ml.log_artifact("ibtracs_dataset", ibtracs_path, is_input=False)
